DAPI-1406: fix no attribute group message when a product evaluation has not been refreshed
