+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/hooks/AttributeGroup/useProductEvaluatedAttributeGroups.ts
@@ -1,10 +1,9 @@
 import {fetchAllAttributeGroupsDqiStatus, fetchAttributeGroupsByCode} from '../../fetcher';
 import {useEffect, useState} from "react";
 import {useCatalogContext, useFetchProductFamilyInformation, useProductEvaluation} from "../index";
-import {Attribute, Evaluation, Family as FamilyInformation, ProductEvaluation} from "../../../domain";
+import {Attribute, Family as FamilyInformation} from "../../../domain";
 import {AttributeGroupCollection} from "@akeneo-pim-community/settings-ui/src/models";
 import {useMountedRef} from "@akeneo-pim-community/settings-ui/src/hooks";
-import {get as _get} from "lodash";
 
 const useProductEvaluatedAttributeGroups = () => {
   const family = useFetchProductFamilyInformation();
@@ -14,20 +13,6 @@ const useProductEvaluatedAttributeGroups = () => {
   const [evaluatedGroups, setEvaluatedGroups] = useState<null | AttributeGroupCollection>(null);
   const [allGroupsEvaluated, setAllGroupsEvaluated] = useState<boolean>(false);
   const mountedRef = useMountedRef();
-
-  const isProductEvaluationPending = (evaluation: ProductEvaluation, channel: string, locale: string) => {
-    if (Object.keys(evaluation).length === 0) {
-      return true;
-    }
-
-    const axisInProgress: any = Object.keys(evaluation).filter((axisCode: string) => {
-      const axisEvaluation: Evaluation = _get(evaluation, [axisCode, channel, locale]);
-
-      return axisEvaluation.rate?.value === null;
-    });
-
-    return axisInProgress.length > 0;
-  }
 
   const extractFamilyAttributeGroupCodes = (family: FamilyInformation) => {
     let familyAttributeGroups = family.attributes.map((attribute: Attribute) => attribute.group);
@@ -41,7 +26,7 @@ const useProductEvaluatedAttributeGroups = () => {
   }
 
   useEffect(() => {
-    if (channel && locale && evaluation && !isProductEvaluationPending(evaluation, channel, locale)) {
+    if (channel && locale && evaluation) {
       (async () => {
         const response = await fetchAllAttributeGroupsDqiStatus();
         if (mountedRef.current) {
@@ -71,7 +56,6 @@ const useProductEvaluatedAttributeGroups = () => {
           setEvaluatedGroups(attributeGroups);
         }
       } else {
-        setEvaluatedGroups({});
         setAllGroupsEvaluated(true);
       }
     })();
